third_party/WebKit/LayoutTests/fast/js/webidl-operation-functions-should-have-no-prototype.html

<!DOCTYPE html>
<script src="../../resources/js-test.js"></script>
<script>

description("WebIDL operation functions should not have a prototype property.");

shouldBeFalse('"prototype" in document.appendChild');

</script>
